Add Home and Map Pages with Embedded Google Maps
* Add Django and Pre-Commit Infra

* Added hompage

We created a new home app which is connected through the localhost. If you go to the root directory of the website, it will run the homeapp which will display the homepage.
@neunis @brforest @dvasavda

* Linked button to Google Maps page

As the customer requested, we were able to link the "Find Foodbanks" button to another app (webapp) which will display the Google Maps embed in the future.
@neunis @brforest @dvasavda

* Change Travis CI Comment

* Linked button to Google Maps page

As the customer requested, we were able to link the "Find Foodbanks" button to another app (webapp) which will display the Google Maps embed in the future.
@neunis @brforest @dvasavda

* Resolving merge conflicts and adding Google Maps embed

* Beautified HTML code

No significant changes were made to the files. No logical changes were made either. Beautified the HTML code to make it more readable, so all of our <> are aligned properly.

* Updated styling

Updated home and map_page styling. Examples include changing color schemes to make it more "clickable" and friendly, changing certain images to maintain relevency, and resizing divs.

* Clean up Map Page