Added place details
This commit adds place details in the table below the map.

Note: this does not show the
details on the map with the icons; therefore this commit incurs a little technical debt, as it will take a little repositioning of code to efficiently use the detailed information both in the table and on the map icons.