Populate table (#64)
* simple table framework

* Change Google Maps JS Loading to API Code (needs API 3 syntax)

* Add button linking to 'events_needs/' on home page

* Delete __init__.cpython-37.pyc

* Delete admin.cpython-37.pyc

* Delete models.cpython-37.pyc

* Delete urls.cpython-37.pyc

* Delete views.cpython-37.pyc

* Delete __init__.cpython-37.pyc

* Add working embedded Google Maps map and search results

* Fix search to return filtered results

This commit adds keyword filtering to the nearby search.  Now the
results are filtered by "food bank" keyword.

* Results display in table

The query results now display in the table below the map.  Currently,
only the name and address of the item display.

* Added place details

This commit adds place details in the table below the map.

Note: this does not show the
details on the map with the icons; therefore this commit incurs a little technical debt, as it will take a little repositioning of code to efficiently use the detailed information both in the table and on the map icons.

* Add click listeners to map icons

Map icons now show information about the item when clicked on.

* Remove unnecessary __pycache__/ files and update .gitignore files